moved all iexcloud code to com.options.iexcloud
